+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     buildToolsVersion "29.0.2"
     defaultConfig {
         applicationId "com.example.gollect"
-        minSdkVersion 28
+        minSdkVersion 23
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
