+++ app/src/main/AndroidManifest.xml
@@ -37,7 +37,7 @@
 <!--            </intent-filter>-->
         </activity>
         <service android:name=".utility.NotificationListener"
-            android:label="My Notification Listener"
+            android:label="Notification Listener"
             android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE">
             <intent-filter>
                 <action android:name="android.service.notification.NotificationListenerService" />
