+++ app/src/main/java/com/example/gollect/Test2Activity.java
@@ -15,20 +15,5 @@ public class Test2Activity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_test);
-
-        if (!permissionGrantred()) {
-            Intent intent = new Intent(
-                    "android.settings.ACTION_NOTIFICATION_LISTENER_SETTINGS");
-            startActivity(intent);
-        }
-    }
-
-    private boolean permissionGrantred() {
-        Set<String> sets = NotificationManagerCompat.getEnabledListenerPackages(this);
-        if (sets != null && sets.contains(getPackageName())) {
-            return true;
-        } else {
-            return false;
-        }
     }
 }
